Create applib.
