+++ CMakeLists.txt
@@ -38,13 +38,12 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -static")
 include_directories(thirdparty/dynamic-segment-tree/include/)
 
 add_subdirectory(thirdparty/dynamic-segment-tree)
+
+include_directories(thirdparty/dynamic-segment-tree/include/)
 add_subdirectory(encoding_lib)
-
-include_directories(encoding_lib/include/
-                    thirdparty/dynamic-segment-tree/include/)
-
-add_library(common STATIC common.hpp common.cpp opt_ostream_ref.hpp opt_ostream_ref.cpp)
-
+include_directories(encoding_lib/include/)
+add_subdirectory(applib)
+include_directories(applib/include)
 add_subdirectory(arithmetic_archiever)
 add_subdirectory(arithmetic_a_archiever)
 add_subdirectory(arithmetic_d_archiever)
