+++ arithmetic_a_archiever/decoder.cpp
@@ -12,7 +12,12 @@
 
 #include <ael/arithmetic_decoder.hpp>
 #include <ael/dictionary/adaptive_a_dictionary.hpp>
-#include "../common.hpp"
+#include <ael/data_parser.hpp>
+#include <ael/byte_data_constructor.hpp>
+
+#include <applib/opt_ostream.hpp>
+#include <applib/file_opener.hpp>
+#include <applib/word_packer.hpp>
 
 namespace bpo = boost::program_options;
