+++ arithmetic_a_archiever/encoder.cpp
@@ -6,7 +6,10 @@
 
 #include <ael/arithmetic_coder.hpp>
 #include <ael/dictionary/adaptive_a_dictionary.hpp>
-#include "../common.hpp"
+
+#include <applib/opt_ostream.hpp>
+#include <applib/ord_and_tail_splitter.hpp>
+#include <applib/file_opener.hpp>
 
 namespace bpo = boost::program_options;
