+++ arithmetic_archiever/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(arithmetic_archiever)
 
 add_executable(archiever_encoder encoder.cpp)
-target_link_libraries(archiever_encoder Boost::program_options common arithmetic-encoding-lib)
+target_link_libraries(archiever_encoder boost_program_options archievers-applib arithmetic-encoding-lib)
 
 add_executable(archiever_decoder decoder.cpp)
-target_link_libraries(archiever_decoder Boost::program_options common arithmetic-encoding-lib)
+target_link_libraries(archiever_decoder boost_program_options archievers-applib arithmetic-encoding-lib)
