+++ arithmetic_d_archiever/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(arithmetic_d_archiever)
 
 add_executable(archiever_d_encoder encoder.cpp)
-target_link_libraries(archiever_d_encoder Boost::program_options arithmetic-encoding-lib common)
+target_link_libraries(archiever_d_encoder boost_program_options archievers-applib arithmetic-encoding-lib)
 
 add_executable(archiever_d_decoder decoder.cpp)
-target_link_libraries(archiever_d_decoder Boost::program_options arithmetic-encoding-lib common)
+target_link_libraries(archiever_d_decoder boost_program_options archievers-applib arithmetic-encoding-lib)
