+++ arithmetic_d_archiever/decoder.cpp
@@ -11,10 +11,12 @@
 
 #include <ael/arithmetic_decoder.hpp>
 #include <ael/dictionary/adaptive_d_dictionary.hpp>
+#include <ael/byte_data_constructor.hpp>
 
-
-#include "../common.hpp"
-#include "../opt_ostream_ref.hpp"
+#include <applib/opt_ostream.hpp>
+#include <applib/ord_and_tail_splitter.hpp>
+#include <applib/file_opener.hpp>
+#include <applib/word_packer.hpp>
 
 namespace bpo = boost::program_options;
