+++ numerical/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(numerical_archiever)
 
 add_executable(numerical_encoder encoder.cpp)
-target_link_libraries(numerical_encoder Boost::program_options arithmetic-encoding-lib common)
+target_link_libraries(numerical_encoder boost_program_options archievers-applib arithmetic-encoding-lib)
 
 add_executable(numerical_decoder decoder.cpp)
-target_link_libraries(numerical_decoder Boost::program_options arithmetic-encoding-lib common)
+target_link_libraries(numerical_decoder boost_program_options archievers-applib arithmetic-encoding-lib)
