+++ ppma_archiever/CMakeLists.txt
@@ -1,7 +1,7 @@
 project(ppma_archiever)
 
 add_executable(ppma_encoder encoder.cpp)
-target_link_libraries(ppma_encoder Boost::program_options common arithmetic-encoding-lib)
+target_link_libraries(ppma_encoder boost_program_options archievers-applib arithmetic-encoding-lib)
 
 add_executable(ppma_decoder decoder.cpp)
-target_link_libraries(ppma_decoder Boost::program_options common arithmetic-encoding-lib)
+target_link_libraries(ppma_decoder boost_program_options archievers-applib arithmetic-encoding-lib)
